requirements.txt
weaviate-client
langchain-text-splitters
requests
python-dotenv
